+++ R/frequencyPolygon.R
@@ -127,7 +127,7 @@ prepanel.default.freqpolygon <- function(
 freqpoly <- function(x, plot = TRUE, ...) {
   h <- hist(x, plot = FALSE, warn.unused = FALSE, ...)
   res <- hist2freqpolygon(h)
-  res$xname = lazyeval::expr_text(x)
+  res$xname = rlang::expr_text(x)
   dots <- list(...)
   # remove from ... things that plot() might not like.
   bad_args <- c("breaks", "include.lowest", "right", "density", "angle", 
@@ -166,7 +166,7 @@ hist2freqpolygon <- function(hist) {
   brk_counts <- head(res$counts, -1) + ratio * diff(res$counts)
   
   res <-
-    data_frame(
+    dplyr::tibble(
       x = c(res$mids, head(tail(res$breaks, -1), -1)), 
       density = c(res$density, brk_density),
       counts = c(res$counts, brk_counts)) %>%