+++ R/pqrdata.R
@@ -55,13 +55,16 @@ qdata <- function( formula, p = seq(0, 1, 0.25), data = NULL, ...) {
  
 qdata_v <- function( x, p=seq(0, 1, 0.25), na.rm=TRUE, ... ) {
   .check_for_quant_input(x)
-  if ( any(p > 1) | any(p < 0) ) 
+  if (any(p > 1) | any(p < 0)) 
     stop("Prob outside of range 0 to 1.  Do you perhaps want pdata?")
-  qs <- quantile(x, probs=p, na.rm=na.rm, ... )
+  qs <- quantile(x, probs = p, na.rm = na.rm, ...)
+  return(qs)
+  
   if (length(p) == 1) {
-    result <- setNames( c(p, qs), c("p","quantile") )
+    # result <- setNames( c(p, qs), c("p","quantile") )
+    result <- data.frame(quantile = qs, p = p)
   } else {
-    result <- data.frame(quantile = qs, p=p)
+    result <- data.frame(quantile = qs, p = p)
   }
   result
 }
@@ -72,7 +75,7 @@ qdata_v <- function( x, p=seq(0, 1, 0.25), na.rm=TRUE, ... ) {
 qdata_f <- aggregatingFunction1(qdata_v, output.multiple = TRUE, na.rm = TRUE)
 
 #' @rdname pqrdata
-#' @return for `cdata`, a named numerical vector or a data frame giving
+#' @return for `cdata`, a data frame giving
 #' upper and lower limits and the central proportion requested
 #' @examples
 #' data(iris)
@@ -102,9 +105,10 @@ cdata_v <- function( x, p = .95, na.rm = TRUE, ... ) {
   lo <- quantile( x, lo_p, na.rm=na.rm, ... )
   hi <- quantile( x, hi_p, na.rm=na.rm, ... )
   if (length(p) == 1) {
-    result <- setNames( c(lo, hi, p), c("low", "hi", "central.p") )
+    # result <- setNames( c(lo, hi, p), c("lower", "upper", "central.p") )
+    result <- data.frame(lower = lo, upper = hi, central.p = p)
   } else {
-    result <- data.frame(low=lo, hi=hi, central.p=p)
+    result <- data.frame(lower = lo, upper = hi, central.p = p)
   }
   return(result)
 }