+++ NEWS.md
@@ -1,4 +1,14 @@
 
+# mosaic 1.6.0
+ 
+ * Changed return type for `qdata()` so that it is always a named vector.
+ * Changed return type for `cdata()` so taht is is always a data frame.  Also changed 
+ names to "lo" and "hi".
+ * Fixed bug in `xpchisq()` caused by introducing explicit arguments and failing
+ to retain `...`.  (Issue #737)
+ * Fixed bug in `xpt()` caused by introducing explicit arguments and failing
+ to handle missing `ncp` correctly.  (Issue #736)
+
 # mosaic 1.5.0
 
  * `googleMap()` has be deprecated due to change in policy at google. Try `leaflet_map()` as an alternative.